+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/writetag/InputWriteTag2.java
@@ -3,7 +3,7 @@
 // Created: 2004
 ////////////////////////////////////////////////////////////////////////////////
 
-package com.puppycrawl.tools.checkstyle.checks.javadoc;
+package com.puppycrawl.tools.checkstyle.checks.javadoc.writetag;
 
 /**
  * @incomplete This enum needs more code...
